Add Maven wrapper
This PR aims to add a wrapper for Maven via `mvn wrapper:wrapper`. The
main goals are:

  - ease of use for new contributors without a local Maven installation:
    just run `./mvnw clean install` (or `.\mvnw clean install` on Windows)
    and you're good to go
  - a fully encapsulated build setup provided by the project - everything
    required is contained within the repository
  - reproducible builds - Maven version can be enforced via a property
    specified in `.mvn/wrapper/maven-wrapper.properties`

If this repo is not allowed to contain binary files (`maven-wrapper.jar`),
the plugin can be configured to use a source distribution instead.

I haven't touched CI build definitions as I'm not familiar with Cirrus CI.